storage ops moved from python to cpp

+++ tests/e2e/feature/nested-mapping.sol
@@ -5,13 +5,13 @@ pragma solidity ^0.8.10;
 
 
 contract NestedMappingTestCase {
-    mapping(uint256 => mapping(uint256 => uint256)) public map;
+    mapping(uint256 => mapping(uint256 => uint32)) public map;
 
-    function mappingRead(uint256 x, uint256 y) external view returns (uint256) {
+    function mappingRead(uint256 x, uint256 y) external view returns (uint32) {
         return map[x][y];
     }
 
-    function mappingWrite(uint256 k1, uint256 k2, uint256 v) external {
+    function mappingWrite(uint256 k1, uint256 k2, uint32 v) external {
         map[k1][k2] = v;
     }
 }
